feat: restrict

+++ sources/routes/data-umum/jenis-kelamin.ts
@@ -1,7 +1,8 @@
 import express, { Router } from "express";
 
 import { headTitle } from ".";
-import { JenisKelamin } from "../../models";
+
+import { Guru, JenisKelamin, Siswa } from "../../models";
 
 export const dataUmumJenisKelaminRouter = Router();
 
@@ -215,18 +216,18 @@ dataUmumJenisKelaminRouter
         const dataExist = await JenisKelamin.exists({ _id: id });
 
         if (dataExist != null) {
-            // const dataIsUsed = await Siswa.exists({ id_jenis_kelamin: id });
+            const dataIsUsed = (await Siswa.exists({ id_jenis_kelamin: id })) && (await Guru.exists({ id_jenis_kelamin: id }));
 
-            // if (dataIsUsed == null) {
-            try {
-                await JenisKelamin.deleteOne({ _id: id });
-                res.redirect("./?response=success");
-            } catch (error) {
-                res.redirect(`delete?id=${id}&response=error`);
+            if (dataIsUsed == null) {
+                try {
+                    await JenisKelamin.deleteOne({ _id: id });
+                    res.redirect("./?response=success");
+                } catch (error) {
+                    res.redirect(`delete?id=${id}&response=error`);
+                }
+            } else if (dataIsUsed != null) {
+                res.redirect(`delete?id=${id}&response=error&text=Data digunakan di data lain`);
             }
-            // } else if (dataIsUsed != null) {
-            //     res.redirect(`delete?id=${id}&response=error&text=Data digunakan di data lain`);
-            // }
         } else if (dataExist == null) {
             res.redirect("./?response=error&text=Data tidak valid");
         }
